update plt result

+++ Project02/my_model.py
@@ -8,14 +8,14 @@ def calc_model(theta_arg, x_in, y_in):
     y = np.array(y_in).reshape((m, 1))
     for ti in range(len(theta_arg)):
         for i in range(len(x)):
-            x[i][ti] = x_in[i]**int(theta_arg[ti]) 
+            x[i][ti] = x_in[i]**float(theta_arg[ti]) 
     theta = np.linalg.pinv(x.T @ x) @ x.T @ y
     return theta
 def predict(theta_arg, theta, x_in):
     y = 0
     theta_arg = theta_arg.split()
     for i in range(len(theta)):
-        y += x_in**int(theta_arg[i]) * theta[i]
+        y += x_in**float(theta_arg[i]) * theta[i]
     return y
 if __name__ == '__main__':
     x = [1, 2, 3, 4, 5]
